fix: move board out of it's unnecessary folder
