+++ src/index.ts
@@ -2,7 +2,7 @@ export { default as RecoverTimeGraph } from './RecoverTimeGraph';
 export { default as ChangeFailureRateGraph } from './ChangeFailureRateGraph';
 export { default as ChangeLeadTimeGraph } from './ChangeLeadTimeGraph';
 export { default as DeploymentFrequencyGraph } from './DeploymentFrequencyGraph';
-export { default as Board } from './Boards/Board';
+export { default as Board } from './Board';
 export { default as TrendGraph } from './TrendGraph';
 export { default as TrendIndicator } from './icons/TrendIndicator';
 import './global.css';
